+++ package.json
@@ -17,6 +17,7 @@
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
     "react-native": "0.76.6",
+    "react-native-stars": "^1.2.2",
     "undici": "^7.2.3"
   },
   "devDependencies": {
